@@ -15,33 +15,38 @@
     "types": "esm/index.d.ts",
     "sideEffects": false,
     "dependencies": {
+        "@emotion/react": "^11.9.3",
+        "@emotion/styled": "^11.9.3",
+        "@mui/icons-material": "^5.8.4",
+        "@mui/material": "^5.8.4",
+        "@mui/styles": "^5.8.4",
         "ra-supabase-core": "^1.2.1"
     },
     "devDependencies": {
-        "@material-ui/core": "^4.11.2",
-        "@material-ui/icons": "^4.11.2",
-        "@material-ui/styles": "^4.11.2",
-        "@supabase/supabase-js": "^1.21.0",
+        "@material-ui/core": "^5.0.0-beta.5",
+        "@material-ui/icons": "^5.0.0-beta.5",
+        "@material-ui/styles": "^5.0.0-beta.5",
+        "@supabase/supabase-js": "^1.35.0",
         "final-form": "^4.20.2",
-        "ra-core": "^3.17.3",
-        "ra-ui-materialui": "^3.17.3",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0",
-        "react-final-form": "^6.5.2",
-        "react-router": "^5.2.0"
+        "ra-core": "^4.1.3",
+        "ra-ui-materialui": "^4.1.3",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
+        "react-final-form": "^6.5.9",
+        "react-router": "^6.3.0"
     },
     "peerDependencies": {
-        "@material-ui/core": "^4.11.2",
-        "@material-ui/icons": "^4.11.2",
-        "@material-ui/styles": "^4.11.2",
-        "@supabase/supabase-js": "^1.21.0",
+        "@material-ui/core": "^5.0.0-beta.5",
+        "@material-ui/icons": "^5.0.0-beta.5",
+        "@material-ui/styles": "^5.0.0-beta.5",
+        "@supabase/supabase-js": "^1.35.0",
         "final-form": "^4.20.2",
-        "ra-core": "^3.17.0",
-        "ra-ui-materialui": "^3.17.0",
-        "react": "^16.9.0 || ^17.0.0",
-        "react-dom": "^16.9.0 || ^17.0.0",
-        "react-final-form": "^6.5.2",
-        "react-router": "^5.1.0"
+        "ra-core": "^4.1.3",
+        "ra-ui-materialui": "4.1.3",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
+        "react-final-form": "^6.5.9-",
+        "react-router": "^6.3.0"
     },
     "scripts": {
         "build": "yarn run build-cjs && yarn run build-esm",
