@@ -38,7 +38,7 @@ import * as React from 'react';
 import { useTranslate } from 'ra-core';
 import { Field, Form } from 'react-final-form';
 import { Button, CardActions, CircularProgress } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@mui/styles';
 import { useSetPassword, useSupabaseAccessToken } from 'ra-supabase-core';
 import { Input } from './Input';
 export var SetPasswordForm = function (props) {
