@@ -1,7 +1,11 @@
 "use strict";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
 }) : (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     o[k2] = m[k];
@@ -60,15 +64,15 @@ var React = __importStar(require("react"));
 var ra_core_1 = require("ra-core");
 var react_final_form_1 = require("react-final-form");
 var core_1 = require("@material-ui/core");
-var styles_1 = require("@material-ui/core/styles");
+var styles_1 = require("@mui/styles");
 var ra_supabase_core_1 = require("ra-supabase-core");
 var Input_1 = require("./Input");
 var SetPasswordForm = function (props) {
     var classes = useStyles(props);
-    var translate = ra_core_1.useTranslate();
-    var access_token = ra_supabase_core_1.useSupabaseAccessToken();
+    var translate = (0, ra_core_1.useTranslate)();
+    var access_token = (0, ra_supabase_core_1.useSupabaseAccessToken)();
     var onSuccess = props.onSuccess, onFailure = props.onFailure;
-    var setPassword = ra_supabase_core_1.useSetPassword({ onSuccess: onSuccess, onFailure: onFailure });
+    var setPassword = (0, ra_supabase_core_1.useSetPassword)({ onSuccess: onSuccess, onFailure: onFailure });
     var validate = function (values) {
         var errors = { email: undefined, password: undefined };
         if (!values.password) {
@@ -109,7 +113,7 @@ var SetPasswordForm = function (props) {
         } }));
 };
 exports.SetPasswordForm = SetPasswordForm;
-var useStyles = styles_1.makeStyles(function (theme) { return ({
+var useStyles = (0, styles_1.makeStyles)(function (theme) { return ({
     form: {
         padding: '0 1em 1em 1em',
     },
