@@ -1,7 +1,11 @@
 "use strict";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
 }) : (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     o[k2] = m[k];
@@ -64,11 +68,11 @@ var useSetPassword_1 = require("./useSetPassword");
 describe('useSetPassword', function () {
     var UseSetPassword = function (_a) {
         var onSuccess = _a.onSuccess, onFailure = _a.onFailure;
-        var setPassword = useSetPassword_1.useSetPassword({
+        var setPassword = (0, useSetPassword_1.useSetPassword)({
             onSuccess: onSuccess,
             onFailure: onFailure,
         });
-        react_1.useEffect(function () {
+        (0, react_1.useEffect)(function () {
             setPassword({ access_token: 'token', password: 'bazinga' });
         }, [setPassword]);
         return null;
@@ -87,13 +91,13 @@ describe('useSetPassword', function () {
                         setPassword: jest.fn().mockResolvedValue(undefined),
                     };
                     myOnSuccess = jest.fn();
-                    ra_test_1.renderWithRedux(React.createElement(ra_core_1.AuthContext.Provider, { value: authProvider },
+                    (0, ra_test_1.renderWithRedux)(React.createElement(ra_core_1.AuthContext.Provider, { value: authProvider },
                         React.createElement(UseSetPassword, { onSuccess: myOnSuccess })));
                     expect(authProvider.setPassword).toHaveBeenCalledWith({
                         access_token: 'token',
                         password: 'bazinga',
                     });
-                    return [4 /*yield*/, react_2.waitFor(function () {
+                    return [4 /*yield*/, (0, react_2.waitFor)(function () {
                             expect(myOnSuccess).toHaveBeenCalledTimes(1);
                         })];
                 case 1:
@@ -117,13 +121,13 @@ describe('useSetPassword', function () {
                         setPassword: jest.fn().mockRejectedValue(error),
                     };
                     myOnFailure = jest.fn();
-                    ra_test_1.renderWithRedux(React.createElement(ra_core_1.AuthContext.Provider, { value: authProvider },
+                    (0, ra_test_1.renderWithRedux)(React.createElement(ra_core_1.AuthContext.Provider, { value: authProvider },
                         React.createElement(UseSetPassword, { onFailure: myOnFailure })));
                     expect(authProvider.setPassword).toHaveBeenCalledWith({
                         access_token: 'token',
                         password: 'bazinga',
                     });
-                    return [4 /*yield*/, react_2.waitFor(function () {
+                    return [4 /*yield*/, (0, react_2.waitFor)(function () {
                             expect(myOnFailure).toHaveBeenCalledWith(error);
                         })];
                 case 1:
