@@ -1,7 +1,11 @@
 "use strict";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
 }) : (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     o[k2] = m[k];
@@ -26,7 +30,7 @@ var AuthLayout_1 = require("./AuthLayout");
 var LoginForm_1 = require("./LoginForm");
 var LoginPage = function (_a) {
     var _b = _a.children, children = _b === void 0 ? React.createElement(LoginForm_1.LoginForm, null) : _b;
-    ra_supabase_core_1.useRedirectIfAuthenticated();
+    (0, ra_supabase_core_1.useRedirectIfAuthenticated)();
     return React.createElement(AuthLayout_1.AuthLayout, null, children);
 };
 exports.LoginPage = LoginPage;
