@@ -1,7 +1,11 @@
 "use strict";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
 }) : (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     o[k2] = m[k];
@@ -22,7 +26,9 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.authRoutes = void 0;
 var React = __importStar(require("react"));
 var ra_core_1 = require("ra-core");
+var react_router_1 = require("react-router");
 var SetPasswordPage_1 = require("./SetPasswordPage");
 exports.authRoutes = [
-    React.createElement(ra_core_1.RouteWithoutLayout, { noLayout: true, path: "/set-password", render: function () { return React.createElement(SetPasswordPage_1.SetPasswordPage, null); } }),
+    React.createElement(ra_core_1.CustomRoutes, { noLayout: true },
+        React.createElement(react_router_1.Route, { element: React.createElement(SetPasswordPage_1.SetPasswordPage, null), path: "/set-password" }))
 ];
